Fixed scatter functionality to use for the user interface. Added properties for the min and max values for both scaling and rotation.
Rearranged when the list of vertices is stored after converting them to vertices under the get_vertices method.

Changed the argument passed in to the move_instances, random_rot and random_scaling method to set values for the correct object.
Added the move, rotation and scaling instances methods to be within a for loop that passes an argument of the source object instance.
Under random_rot the variable name that determined a random value was changed from random_rot to random_seed.