Removed random.seed placed at the beginning so that vertices can be found randomly each time the program is run. Added selection_percentage variable. Is set in the constructor and in _init_from_objects method. Created the _set_sel_percentage_sbx_properties_ui method to set the selection percentage properties. Created the _create_percent_sel_layout, _create_sel_percent_headers and _create_selected_percent_sbx to set up the user interface.
The is_aligned variable was created to check if the user wants to constrain the source instances normals to the destination object.
An if statement is added to to the move_instances method to align the source instances if the user determines it.
Created _constrain_normals method, currently does nothing but, will constrain the normals of the source instances.